let Slick store use queries from JDBC store
