+++ akka-projection-slick/src/main/resources/reference.conf
@@ -35,5 +35,7 @@ akka.projection.slick {
     # Oracle schema is case sensitive and is defined with uppercase, this property is therefore ignore when using Oracle
     use-lowercase-schema = true
   }
+
+  debug.verbose-offset-store-logging = false
 }
 //#config
